Enable loading of .archrc from project roots using rc module
