+++ lib/config.js
@@ -1,0 +1,6 @@
+(function(){
+  var rc, defaultConfig;
+  rc = require('rc');
+  defaultConfig = require('./default-config');
+  module.exports = rc('arch', defaultConfig);
+}).call(this);
