+++ lib/server/server.js
@@ -1,5 +1,5 @@
 (function(){
-  var express, path, bluebird, bodyParser, bundler, livescript, register, renderBody, ref$, each, values, filter, find, flatten, map, first, defaults, archGet, archPost;
+  var express, path, bluebird, bodyParser, bundler, livescript, register, paths, getConfig, renderBody, ref$, each, values, filter, find, flatten, map, first, archGet, archPost, __template;
   express = require('express');
   path = require('path');
   bluebird = require('bluebird');
@@ -7,27 +7,15 @@
   bundler = require('../bundler');
   livescript = require('livescript');
   register = require('babel/register');
+  paths = require('./paths');
+  getConfig = require('./get-config');
   renderBody = require('./render').renderBody;
   ref$ = require('prelude-ls'), each = ref$.each, values = ref$.values, filter = ref$.filter, find = ref$.find, flatten = ref$.flatten, map = ref$.map, first = ref$.first;
-  defaults = {
-    environment: process.env.NODE_ENV || 'development',
-    port: process.env.ARCH_PORT || 3000,
-    paths: {
-      app: {
-        abs: path.resolve('.'),
-        rel: path.relative(__dirname, path.resolve('.'))
-      },
-      arch: {
-        abs: path.dirname(require.resolve("../../package.json")),
-        rel: path.relative(path.resolve('.'), path.dirname(require.resolve("../../package.json")))
-      },
-      'public': 'dist'
-    }
-  };
-  module.exports = function(options){
-    var app, get, post;
-    options = import$(clone$(defaults), options);
-    app = options.app || require(options.paths.app.rel);
+  module.exports = function(opts){
+    var options, app, get, post;
+    opts == null && (opts = {});
+    options = getConfig(opts);
+    app = require(paths.app.rel);
     get = function(req, res){
       console.log("GET", req.originalUrl);
       return archGet(app, req.originalUrl, options).spread(function(status, headers, body){
@@ -47,10 +35,10 @@
     return {
       start: function(cb){
         var server, listener;
-        server = express().use("/" + options.paths['public'], express['static'](path.join(options.paths.app.abs, options.paths['public']))).use(bodyParser.urlencoded({
+        server = express().use("/" + options['public'], express['static'](path.join(paths.app.abs, options['public']))).use(bodyParser.urlencoded({
           extended: false
         })).get('*', get).post('*', post);
-        bundler.bundle(options.paths, options.environment === 'development', function(ids){
+        bundler.bundle(paths, options.environment === 'development', function(ids){
           var done, id, parents, e;
           done = [];
           while (id = first(ids)) {
@@ -70,7 +58,7 @@
           })(
           done);
           try {
-            return app = require(options.paths.app.rel);
+            return app = require(paths.app.rel);
           } catch (e$) {
             e = e$;
             return console.error('Error in changed files when restarting server');
@@ -132,15 +120,7 @@
       return [200, {}, html];
     });
   };
-  function import$(obj, src){
-    var own = {}.hasOwnProperty;
-    for (var key in src) if (own.call(src, key)) obj[key] = src[key];
-    return obj;
-  }
-  function clone$(it){
-    function fun(){} fun.prototype = it;
-    return new fun;
-  }
+  __template = jade.compileFile(path.join(__dirname, 'index.jade'));
   function in$(x, xs){
     var i = -1, l = xs.length >>> 0;
     while (++i < l) if (x === xs[i]) return true;
