+++ package.json
@@ -60,6 +60,7 @@
     "gulp-livescript": "^2.4.0",
     "gulp-load-plugins": "^1.0.0-rc.1",
     "gulp-strip-code": "^0.1.2",
-    "jasmine": "^2.3.1"
+    "jasmine": "^2.3.1",
+    "rc": "^1.0.3"
   }
 }
