+++ src/get-config.ls
@@ -1,0 +1,3 @@
+require! <[ rc ./default-config ]>
+
+module.exports = (opts = {}) -> rc 'arch', ^^default-config import opts
